Revert "Refactor evaluators into separate package"
This reverts commit 55e4086ce924ad7e331b97d2e9eafaa7ab4bba2c.
