+++ core/src/main/scala/fortress/operations/EvaluationInliner.scala
