+++ core/src/main/scala/fortress/operations/Evaluator.scala
