--- core/src/main/scala/fortress/operations/Simplification/DefinitionJit.scala
@@ -1,10 +1,0 @@
-package fortress.operations
-
-import fortress.msfol._
-
-/** Compiles function definitions to Scala functions on demand. */
-class DefinitionJit(theory: Theory) {
-
-
-
-}
