+++ core/src/main/scala/fortress/transformers/EvaluateTransformer.scala
@@ -1,6 +1,6 @@
 package fortress.transformers
 
-import fortress.operations.Simplification.EvaluationInliner
+import fortress.operations.EvaluationInliner
 import fortress.problemstate.ProblemState
 
 /**
